+++ src/core.ts
@@ -1,11 +1,13 @@
-export * from './aloneFigure';
 export * from './axes';
+export * from "./base.shapes";
 export * from "./collections";
 export * from "./colors";
 export * from "./constants";
 export * from './figures';
+export * from "./functions.shapes";
 export * from "./functions";
 export * from "./multiplot";
-export * from "./shapedObjects";
-export * from "./shapes";
+export * from "./primitives.shapes";
+export * from './remoteFigure';
+export * from "./shapes.shapes";
 export * from "./styles";
